refactors import and export statements of moved components
